requirements.txt
# tensorflow[and-cuda]
tensorflow-probability==0.24.0
gpflow
matplotlib
tqdm
